misuse of window.onload
window.onload can actually cause problems with other event listeners.

The common pratice is to use an event listener & then possibly use window.onload as a fallback for old or non-mainstream browsers.

JQuery's event listener function basically does all of that for you
(the Vanilla version of this is shown in my last PR).