various edits
- added previous edit back that was omitted in error
- more DOM checks prior to manipulating nodes

I found some other errors while in the admin for the active block image toggle.

After these edits there are no errors in the block admin and the toggle works as expected.

Just a note that the 2 image color choices are not really visibly different to my eyes because I'm red-green color blind. Has anyone else ever mentioned this before?  When I toggle it the image is changing (I can see it in the console) but to my eyes there isn't much of a noticeable difference.